refactor: fold changedFilesSince into libs/utils/git.ts
changed-files.ts was a generic git-exec wrapper with no kg-domain knowledge,
near-identical to utils/git.ts. Merge it in — both now share one git() exec
helper — and drop the misplaced kg-domain file.
