+++ libs/knowledge-graph/service.ts
@@ -14,10 +14,9 @@ import { CodeAnchorResolver } from "./code-anchors/resolver.js";
 import { hashAnchorSpan, statAnchorFile } from "./code-anchors/span-hash.js";
 import { classifyFreshness, hasContentDrift, type AnchorCheck, type FreshnessVerdict } from "./code-anchors/freshness.js";
 import { freshnessChecks, indexFingerprintsByClaim } from "./anchor-fingerprints.js";
-import { changedFilesSince } from "./changed-files.js";
 import { buildAnchorInvalidation, type ClaimDemotion } from "./anchor-invalidation.js";
 import type { ResolvedCodeAnchorStatus } from "./code-anchors/types.js";
-import { gitHeadSha } from "../utils/git.js";
+import { changedFilesSince, gitHeadSha } from "../utils/git.js";
 import { defaultDatabasePath, openDatabase } from "../storage/sqlite/db.js";
 import type { AnchorFingerprintInput, SqliteRepository } from "../storage/sqlite/repository.js";
 import { SqliteRepository as SqliteKnowledgeGraphRepository } from "../storage/sqlite/repository.js";
